+++ src/lib/components/space/info/SessionInfo.svelte
