+++ src/lib/components/space/members/SessionMemberAvatar.svelte
