+++ src/lib/components/space/search/EmptySearchResultIndicator.svelte
