+++ src/lib/components/space/track/MusicSearchResult.svelte
