+++ src/routes/session/[sessionId]/+page.svelte
@@ -2,16 +2,16 @@
 	import { onDestroy, onMount } from 'svelte';
 	import { currentSession } from '$stores/session';
 	import { spaceMemberStore, spaceRoleStore, spaceStore } from '$stores/space/index';
-	import TrackSearchTab from '$components/music/TrackSearchTab.svelte';
-	import TrackQueueList from '$components/music/TrackQueueList.svelte';
-	import MusicPlayerController from '$components/music/MusicPlayerController.svelte';
-	import SessionInfo from '$components/music/SessionInfo.svelte';
+	import TrackSearchTab from '$components/space/track/TrackSearchTab.svelte';
+	import TrackQueueList from '$components/space/queue/TrackQueueList.svelte';
+	import MusicPlayerController from '$components/space/player/MusicPlayerController.svelte';
+	import SessionInfo from '$components/space/info/SessionInfo.svelte';
 	import { userConfigStore, userStore } from '$stores/auth/user';
 	import seo from '$constants/seo';
 	import type { RealtimeChannel } from '@supabase/supabase-js';
 	import { Drawer, Modal } from 'flowbite-svelte';
 	import SessionHelpModal from '$components/modals/SessionHelpModal.svelte';
-	import SpaceMembers from '$components/music/members/SpaceMembers.svelte';
+	import SpaceMembers from '$components/space/members/SpaceMembers.svelte';
 	import { sineIn } from 'svelte/easing';
 	import Icon from '@iconify/svelte';
 	import type { SessionMember } from '$lib/types/session/member.interface.js';
